Add files via upload
This project shows if the climate is important about the gdp's of the countries. Most important thing here is, as in everything you always need to look for the graphs. The best way to understand a dataset is to visualization. Yes descriptive statistics always ready to help but for non-technical audience and also for the sake of our understanding best thing to do is always visualazing. Do not forget that when you're analyzing data.