Add files via upload
In this code you can see that it is related to the article "I just ran 2 Million Regressions" by Xalai Martin. Dependent variable here is the growth and we have a bunch of independent explanatory variables. Here we try to understand what are the things that affects growth rate. We have 3 main explanatory variable. It is an important point because as we start anaylzing we have always some kind of variables that we can predict the signs and relations. For example if you're looking for the relation between wage and education you would wait that education  should have a positive sign, so positive correlation. If there is negative sign then you can think of there is omitted variable biaa, simultaneouty bias or wrong functional form. This may caused by the low number of the observations too. So to sum up it is important to have control variables to understand whats going on the data by looking regression and also you should always look for graphs especially scatter plots to understand data in a better way.